backend/requirements.txt
fastapi
uvicorn
openai-whisper
spacy
google-generativeai
gTTS
python-multipart
scikit-learn
joblib
pandas
numpy
